requirements.txt
autopep8
beautifulsoup4
certifi
chardet
comtypes
idna
MouseInfo
Pillow
PyAudio
PyAutoGUI
pycodestyle
PyGetWindow
PyMsgBox
pyperclip
# pypiwin32
PyRect
PyScreeze
python-decouple
pyttsx3
PyTweening
pywhatkit
#pywin32==301
requests
soupsieve
SpeechRecognition
toml
urllib3
wikipedia
openpyxl